Turn on core dumps on linux for the inner workbench process
Has to switch from multiprocessing to subprocess in order to do this (needed the preexec option).
Through some testing, this should preserve the desired spawn behaviour.
A new module 'workbench_process.py' has been created to use with subprocess.run (can't pass it a function)
